Исправляет ошибки(отображение при переключение фильтра)
