+++ js/main.js
@@ -1,8 +1,7 @@
 import './form.js';
 import {renderGallery} from './full-photo.js';
 import { getData } from './api.js';
-import {showAlert} from './message.js';
-import {debounce} from './util.js';
+import {debounce, showAlert} from './util.js';
 import {init, getFilterPictures} from './photo-filters.js';
 
 const debounceRenderGallery = debounce(renderGallery);
