+++ js/util.js
@@ -1,4 +1,26 @@
+const ALERT_SHOW_TIME = 5000;
+
 const isEscapeKey = (evt) => evt.key === 'Escape';
+
+const showAlert = (message) => {
+  const alertContainer = document.createElement('div');
+  alertContainer.style.zIndex = '100';
+  alertContainer.style.position = 'absolute';
+  alertContainer.style.left = '120px';
+  alertContainer.style.top = '0';
+  alertContainer.style.right = '120px';
+  alertContainer.style.padding = '10px 3px';
+  alertContainer.style.fontSize = '20px';
+  alertContainer.style.textAlign = 'center';
+  alertContainer.style.backgroundColor = 'red';
+  alertContainer.textContent = message;
+
+  document.body.append(alertContainer);
+
+  setTimeout(() => {
+    alertContainer.remove();
+  }, ALERT_SHOW_TIME);
+};
 
 const debounce = (callback, timeoutDelay = 500) => {
   let timeoutId;
@@ -8,4 +30,4 @@ const debounce = (callback, timeoutDelay = 500) => {
   };
 };
 
-export {isEscapeKey, debounce};
+export {isEscapeKey, showAlert, debounce};
